requirements.txt
flask~=2.2.2
flask-sqlalchemy~=3.0.2
flask-wtf~=1.0.1
flask-login~=0.6.2
email_validator~=1.1.3
mammoth~=1.4.15
flask-restful~=0.3.8